bot/pyproject.toml
[project]
name = "vsb-bot"
version = "0.0.1"
dependencies = [
    "aiohappyeyeballs==2.4.4",
    "aiohttp==3.11.11",
    "aiosignal==1.3.2",
    "certifi==2024.12.14",
    "charset-normalizer==3.4.1",
    "discord.py==2.4.0",
    "frozenlist==1.5.0",
    "idna==3.10",
    "multidict==6.1.0",
    "propcache==0.2.1",
    "requests==2.32.3",
    "six==1.17.0",
    "urllib3==2.3.0",
    "yarl==1.18.3",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = ["*.tests", "*.tests.*", "tests.*", "tests"]

[tool.setuptools]
include-package-data = true
